Moving to unique_ptr where possible and moving the thermal operator to a separate file
