+++ src/solvers/CMakeLists.txt
@@ -9,6 +9,7 @@ set(solvers_sources
     nonlinear_solid_solver.cpp
     nonlinear_solid_operators.cpp
     thermal_solver.cpp
+    thermal_operators.cpp
     elasticity_solver.cpp
     )
 
@@ -17,6 +18,7 @@ set(solvers_headers
     nonlinear_solid_solver.hpp
     nonlinear_solid_operators.hpp
     thermal_solver.hpp
+    thermal_operators.hpp
     elasticity_solver.hpp
     )
